requirements.txt
sqlalchemy>=2.0.0
databricks-sql-connector>=3.1.0
psycopg2-binary>=2.9.9
pyodbc>=5.1.0
python-dotenv>=1.0.1
fastmcp>=0.3.0
pandas>=2.2.2

